corrected empty space in return actuation
the empty space that continues in the return of the devices performance has been fixed.
And also in the docs/apiary.apib files there was this same space, and I removed it.
And in the test files "test_device_handler.py" there was also, but I have already removed those spaces mentioned above.